requirements.txt
Appium-Python-Client=>5.2.5
pytest=>9.0.2